+++ package.json
@@ -16,6 +16,8 @@
     "@ckeditor/ckeditor5-build-classic": "^34.2.0",
     "@ckeditor/ckeditor5-dev-webpack-plugin": "^30.3.5",
     "@ckeditor/ckeditor5-vue2": "^3.0.1",
+    "@nuxtjs/axios": "^5.13.6",
+    "@nuxtjs/proxy": "^2.1.0",
     "axios": "^0.27.2",
     "core-js": "^3.19.3",
     "nuxt": "^2.15.8",
